added a new function to export ply meshes from a stack - this is to be implemented in trabmap_map function as soon as I fix a big that ends up flipping the .ply mesh.
Edited an error in the pointcloud step of the trabmap_map function to account for the changes to the coordinate system associated with the padding. this is simply done by subtracting voi_interval (which is used for padding) from the x,y, and z coordinates.

This makes sure the meshes and pointclouds overlap in space which is important for statistical analyses